Added product images
Added product images. The product table now hold a row for ImagePath.
which points to the location of the product images in the project.
"./backend/Images/Products/". When getting products use this path to add
the images to the response from the API. And when creating products save
the uploaded image to that location, add the path to the product-entity
in the db.